[refactor] optimization file name and directory structure on pr review.
