+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { RouterProvider } from "react-router-dom";
-import router from "./authentication/router";
+import router from "./routes";
 import { ToastContainer } from "react-toastify";
 import React from "react";
