+++ src/components/Auth/Login.tsx
@@ -12,8 +12,8 @@ import { AxiosResponse } from "axios";
 import { Link, useNavigate } from "react-router-dom";
 import { toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-import bg1 from "../../images/bg_left_auth_page.png";
-import bg2 from "../../images/bg_right_auth_page.png";
+import bg1 from "../../images/auth-left-img.png";
+import bg2 from "../../images/auth-right-img.png";
 import FacebookLogin, {
   SuccessResponse,
 } from "@greatsumini/react-facebook-login";
