+++ src/config/axios-interceptor.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
-import { LOCALSTORAGE, ROUTES, ROUTING } from "./constants";
-import { getUserSessionData, ISessionData } from "./user.sessiondata";
+import { LOCALSTORAGE, ROUTES, ROUTING } from "../utils/constants";
+import { getUserSessionData, ISessionData } from "../utils";
 
 export const axiosInstance = axios.create({
   baseURL: process.env.REACT_APP_API_BASE_URL!,
