+++ src/store/index.ts
@@ -1,7 +1,7 @@
 import { useDispatch, useSelector } from "react-redux";
 import { configureStore } from "@reduxjs/toolkit";
 import panelReducer from "../slices/panel/panel.slice";
-import profileReducer from "../slices/profile/profile-slice";
+import profileReducer from "../slices/profile/profile.slice";
 import authReducer from "../slices/auth/auth.slice";
 
 export const store = configureStore({
